Update 2
- Training script now works
- The model isn't available right now
