+++ requirements.txt
@@ -1,1 +1,6 @@
+# torch>=2.0.0,<2.3.0
+numpy>=1.23.0,<1.27.0
+polars>=0.19.0,<0.20.0
+scikit-learn>=1.2.0,<1.5.0
+
