+++ src/recommind_packages/__init__.py
@@ -1,0 +1,5 @@
+from .data.processing import Processor
+from .data.treatment import data_treatment
+from .models.model import NeuMF
+from .trainer import trainer
+from .evaluator_metrics import evaluate_batch_precision_recall
